tec: Move Bookmarks controller
It makes more sense outside of the collections\ namespace since it is
outside of collections page in the application. Also it's more logical
to rename the show action in index (i.e. we list the bookmarks).
